+++ api/filters/authn/authn_test.go
@@ -27,8 +27,8 @@ import (
 	"github.com/Peripli/service-manager/pkg/util"
 	"github.com/Peripli/service-manager/pkg/web"
 	"github.com/Peripli/service-manager/pkg/web/webfakes"
-	"github.com/Peripli/service-manager/security"
-	"github.com/Peripli/service-manager/security/securityfakes"
+	"github.com/Peripli/service-manager/pkg/security"
+	"github.com/Peripli/service-manager/pkg/security/securityfakes"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
@@ -40,7 +40,7 @@ func TestHandler(t *testing.T) {
 
 type testStructure struct {
 	authnResp                *web.User
-	authnDecision            security.AuthenticationDecision
+	authnDecision            security.Decision
 	authnErr                 error
 	request                  *web.Request
 	actualHandlerInvokations int
