+++ pkg/security/interfaces.go
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-// Package security contains logic for setting SM security
 package security
 
 import (
@@ -24,35 +23,13 @@ import (
 	"github.com/Peripli/service-manager/pkg/web"
 )
 
-// AuthenticationDecision represents a Authenticator decision to allow or deny authentication or to abstain from
-// taking a decision
-type AuthenticationDecision int
-
-var decisions = []string{"Allow", "Deny", "Abstain"}
-
-const (
-	// Allow represents an authentication decision to allow to proceed
-	Allow AuthenticationDecision = iota
-
-	// Deny represents an authentication decision to deny proceeding
-	Deny
-
-	// Abstain represents an authentication decision to abstain from deciding - let another component to decide
-	Abstain
-)
-
-// String implements Stringer and converts the decision to human-readable value
-func (a AuthenticationDecision) String() string {
-	return decisions[a]
-}
-
 // Authenticator extracts the authenticator information from the request and
 // returns information about the current user or an error if security was not successful
 //go:generate counterfeiter . Authenticator
 type Authenticator interface {
 	// Authenticate returns information about the user if security is successful, a bool specifying
 	// whether the authenticator ran or not and an error if one occurs
-	Authenticate(req *http.Request) (*web.User, AuthenticationDecision, error)
+	Authenticate(req *http.Request) (*web.User, Decision, error)
 }
 
 // TokenVerifier attempts to verify a token and returns it or an error if the verification was not successful
