+++ pkg/security/securityfakes/fake_authenticator.go
@@ -5,31 +5,31 @@ import (
 	"net/http"
 	"sync"
 
+	"github.com/Peripli/service-manager/pkg/security"
 	"github.com/Peripli/service-manager/pkg/web"
-	"github.com/Peripli/service-manager/security"
 )
 
 type FakeAuthenticator struct {
-	AuthenticateStub        func(req *http.Request) (*web.User, security.AuthenticationDecision, error)
+	AuthenticateStub        func(req *http.Request) (*web.User, security.Decision, error)
 	authenticateMutex       sync.RWMutex
 	authenticateArgsForCall []struct {
 		req *http.Request
 	}
 	authenticateReturns struct {
 		result1 *web.User
-		result2 security.AuthenticationDecision
+		result2 security.Decision
 		result3 error
 	}
 	authenticateReturnsOnCall map[int]struct {
 		result1 *web.User
-		result2 security.AuthenticationDecision
+		result2 security.Decision
 		result3 error
 	}
 	invocations      map[string][][]interface{}
 	invocationsMutex sync.RWMutex
 }
 
-func (fake *FakeAuthenticator) Authenticate(req *http.Request) (*web.User, security.AuthenticationDecision, error) {
+func (fake *FakeAuthenticator) Authenticate(req *http.Request) (*web.User, security.Decision, error) {
 	fake.authenticateMutex.Lock()
 	ret, specificReturn := fake.authenticateReturnsOnCall[len(fake.authenticateArgsForCall)]
 	fake.authenticateArgsForCall = append(fake.authenticateArgsForCall, struct {
@@ -58,27 +58,27 @@ func (fake *FakeAuthenticator) AuthenticateArgsForCall(i int) *http.Request {
 	return fake.authenticateArgsForCall[i].req
 }
 
-func (fake *FakeAuthenticator) AuthenticateReturns(result1 *web.User, result2 security.AuthenticationDecision, result3 error) {
+func (fake *FakeAuthenticator) AuthenticateReturns(result1 *web.User, result2 security.Decision, result3 error) {
 	fake.AuthenticateStub = nil
 	fake.authenticateReturns = struct {
 		result1 *web.User
-		result2 security.AuthenticationDecision
+		result2 security.Decision
 		result3 error
 	}{result1, result2, result3}
 }
 
-func (fake *FakeAuthenticator) AuthenticateReturnsOnCall(i int, result1 *web.User, result2 security.AuthenticationDecision, result3 error) {
+func (fake *FakeAuthenticator) AuthenticateReturnsOnCall(i int, result1 *web.User, result2 security.Decision, result3 error) {
 	fake.AuthenticateStub = nil
 	if fake.authenticateReturnsOnCall == nil {
 		fake.authenticateReturnsOnCall = make(map[int]struct {
 			result1 *web.User
-			result2 security.AuthenticationDecision
+			result2 security.Decision
 			result3 error
 		})
 	}
 	fake.authenticateReturnsOnCall[i] = struct {
 		result1 *web.User
-		result2 security.AuthenticationDecision
+		result2 security.Decision
 		result3 error
 	}{result1, result2, result3}
 }
