+++ pkg/security/securityfakes/fake_key_fetcher.go
@@ -5,7 +5,7 @@ import (
 	"context"
 	"sync"
 
-	"github.com/Peripli/service-manager/security"
+	"github.com/Peripli/service-manager/pkg/security"
 )
 
 type FakeKeyFetcher struct {
