+++ storage/postgres/security.go
@@ -23,7 +23,7 @@ import (
 	"time"
 
 	"github.com/Peripli/service-manager/pkg/log"
-	"github.com/Peripli/service-manager/security"
+	"github.com/Peripli/service-manager/pkg/security"
 	"github.com/jmoiron/sqlx"
 	)
