+++ test/security_test/two_layer_encrypter_test.go
@@ -23,8 +23,8 @@ import (
 	"log"
 	"testing"
 
-	"github.com/Peripli/service-manager/security"
-	"github.com/Peripli/service-manager/security/securityfakes"
+	"github.com/Peripli/service-manager/pkg/security"
+	"github.com/Peripli/service-manager/pkg/security/securityfakes"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
